chore: remove rollup and builds
