--- .babelrc.json
@@ -1,3 +1,0 @@
-{
-    "presets": ["@babel/env"]
-}
